Removes duplicate forenames
Looks like the forename list contained a bunch of duplicates.

It seems to me that most/all duplicates are just missing
accents and diacritics, e.g. "José" -> "Jose". Probably the
list was converted from whatever encoding it was using to
ASCII and lost those symbols in the way.

Still, it's just easier to delete the duplicates than to try
to find out the correct spelling of the original names.